backend/requirements.txt
fastapi>=0.111.0
sqlmodel>=0.0.19
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
boto3>=1.34.0
python-dotenv>=1.0.0
modal>=1.0.0
